Refactor: Decompose useFarmhand into Navigation, Network, DayAdvancement, and Notifications sub-hooks
